# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Clipper",
    "description": "Paste a YouTube URL, choose start/end times and a clip style, and download the trimmed clip. Personal-scale tool for free-tier hosting. NOTE: downloading YouTube content via third-party services may violate YouTube's ToS \u2014 you are responsible for having the rights to the content you clip.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/previews": {
      "post": {
        "tags": [
          "previews"
        ],
        "summary": "Create Preview",
        "description": "Queue a source-video preview (instant stream + background download).\n\nThe frontend polls GET /api/previews/{id} and renders the player +\ntimeline as soon as ``video_url``/``duration`` appear. Asking again for\nthe same URL returns the existing usable preview (ready or in-flight)\ninstead of starting a second download.",
        "operationId": "create_preview_api_previews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/previews/{preview_id}": {
      "get": {
        "tags": [
          "previews"
        ],
        "summary": "Get Preview",
        "operationId": "get_preview_api_previews__preview_id__get",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/previews/{preview_id}/thumbs/{name}": {
      "get": {
        "tags": [
          "previews"
        ],
        "summary": "Preview Thumb",
        "operationId": "preview_thumb_api_previews__preview_id__thumbs__name__get",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preview Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/previews/{preview_id}/video": {
      "get": {
        "tags": [
          "previews"
        ],
        "summary": "Preview Video",
        "description": "Legacy alias: stream the downloaded source file (READY previews only).\n\nRange requests are handled by the file response for video seeking.",
        "operationId": "preview_video_api_previews__preview_id__video_get",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/previews/{preview_id}/stream": {
      "get": {
        "tags": [
          "previews"
        ],
        "summary": "Preview Stream",
        "description": "The preview's single playback URL for its entire lifetime.\n\n- Local file on disk (processing/ready) \u2192 served directly; FileResponse\n  handles Range requests (video seeking).\n- Still streaming from the provider \u2192 the stored, provider-resolved URL\n  is proxied with the browser's Range header forwarded so the <video>\n  element can seek before the download finishes.",
        "operationId": "preview_stream_api_previews__preview_id__stream_get",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePreviewRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreatePreviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "stream_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Provider"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "thumb_count": {
            "type": "integer",
            "title": "Thumb Count"
          },
          "thumbs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Thumbs"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "video_id",
          "status",
          "error",
          "provider",
          "stream_provider",
          "title",
          "duration",
          "width",
          "height",
          "thumb_count",
          "thumbs",
          "video_url",
          "progress",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "title": "PreviewOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}